Call qnnpack's conv setup only if input pointer has changed. (#42008)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/42008

With caching allocator we have increased the likelihood of getting the
same input pointer. With that we can cache qnnpack operator and input
pointer and check if the input pointer is the same. If so we can skip
setup step.

Test Plan:
Ran one of the quantized models to observe
1. No pagefaults due to indirection buffer reallocation.
2. Much less time spent in indirection buffer population.

Imported from OSS

Reviewed By: AshkanAliabadi

Differential Revision: D22726973

fbshipit-source-id: 2dd2a6a6ecf1b5cfa7dde65e384b36a6eab052d7